gateway: split stream-event contracts and the dialect dispatch seam for the module line budget
contracts.py (1015) and streaming_requests.py (1024) crossed the 1000-line
gate. Two mechanical cohesive-section moves, zero behavior change:

- The stream event, usage, and failure contracts (GatewayUsage,
  GatewayEventKind, GatewayEvent, GatewayFailureClass, GatewayFailure)
  move byte-for-byte from gateway/contracts.py to gateway/stream_events.py,
  mirroring the existing reasoning_blocks split; contracts.py re-exports
  every moved name so no caller changes.
- The per-dialect wire dispatch (dialect_stream_payload,
  _fireworks_continuation_required, SERVICE_TIER_DIALECTS) moves
  byte-for-byte from providers/streaming_requests.py to
  providers/dialect_dispatch.py, mirroring the messages_payloads /
  openai_payloads splits; streaming_requests re-exports the dispatch seam
  and the payload builders it aggregated so no caller changes.
