+++ exp/runtime/gateway/stream_events_test.py
@@ -1,0 +1,1 @@
+"""The stream event, usage, and failure contracts are exercised in contracts_test.py."""
